rcarvin: reset the occupied buffers counter in remove_device()
The occupied buffers counter 'mb_cnt' could be incremented when
a buffer would be set up to be filled in the irq handler, and
decremented when one of the buffers would be released in
videobuf_release().

Unfortunatery videobuf_release() invocation could be skipped
if a user application is terminated.
This change must be required for such above situation.