rcarvin: Add "stopping capture" state to buffer release processing V2
This patch adds "stopping capture" status to VIN driver.
Without this patch, the driver can only handle "start" or "stop"
status. But when continuous capturing is used, the actual capture
cannot be stopped at the instant the driver tries to stop it.
So the driver needs one more "stopping" status.
Also the driver doesn't need both of the "disable_on" or "first_op"
states to manage the capturing status.
This patch merges those flags to simplfy processing.

Without this patch, the following issue can occur.
In some cases, when capturing stops, all of the buffers are released
before last the interrupt occurs. As a result, this last interrupt
causes a NULL pointer access in the irq handler.
After applying this patch, those buffers are released after the last
interrupt occurs.