rcarvin: ease the minimal number of buffers required for starting/resuming the continuous transfer
This changes the minimal number of buffers required for starting/resuming
the continuous transfer from four to three. The change could have effect on
the issue that some of applications, such as gstreamer, may not always push
all of buffers which prepared by VIDIOC_REQBUFS.

Notice, the minimal number of buffers required for ENABLING the continuous
transfer has not been changed. The continuous transfer could be enabled
only if four or more buffers are prepared by VIDIOC_REQBUFS.