rcarvin: Adjust field/frame complete interrupt settings in progressive mode.
When progressive and continuous capturing mode is specified,the VIN
will capture an entire frame within each of the odd/even fields that
it returns.
Since the driver interprets every interrupt from the VIN as a
"frame complete" signal, the VIN should send an interrupt to the
driver for each field captured since in this case each "field" is
actually a frame.