video: rcarfb: revise the id string in the fb_fix_screeninfo structure
The current implementation of the id string in the fb_fix_screeninfo
structure has two issues.
The first issue is memory leak. The string may be 17 bytes since
it consists of a 15-byte constant string, an index character, and
a NULL termination.
The second issue is index mismatch between plane and framebuffer.
It means that one 32-bit depth framebuffer may be realized with
two planes.
This change makes the id string fit into the 16-bytes array
and present index(s) of used plane rather than framebuffer index.